Refactored code in aprgprocessfiles module

+++ AprgProcessFiles/AprgProcessFiles/src/ProcessFiles/ReplaceStringInFiles/ReplaceStringInFiles.cpp
@@ -175,20 +175,18 @@ string ReplaceStringInFiles::constructCPlusPlusPrint(
     string result(newPrintStream);
     bool isPercentEncountered(false);
     bool isOnStringLiteral(false);
-    unsigned int printParameterIndex = 0;
+    int printParameterIndex = 0;
     for (char c : printString) {
         bool isParameterAppended(false);
         if (isPercentEncountered) {
             if (isLetter(c)) {
-                if (printParameterIndex < printParameters.size()) {
+                if (printParameterIndex < static_cast<int>(printParameters.size())) {
                     appendParameterToResult(result, isOnStringLiteral, printParameters[printParameterIndex++]);
                     isParameterAppended = true;
-                }
-            } else if (!isNumber(c)) {
+                }            } else if (!isNumber(c)) {
                 appendCharacterToResult(result, isOnStringLiteral, '%');
             }
-        }
-        isPercentEncountered = ('%' == c) || (isNumber(c) && isPercentEncountered);
+        }        isPercentEncountered = ('%' == c) || (isNumber(c) && isPercentEncountered);
         if (!isPercentEncountered && !isParameterAppended) {
             appendCharacterToResult(result, isOnStringLiteral, c);
         }
